authentication-login

+++ app/category/controller.js
@@ -6,6 +6,9 @@ const config = require('../config');
 const fs = require('fs');
 const path = require('path');
 
+// Hak Akses
+const { policyFor } = require('../policy');
+
 async function index(req, res, next) {
     try {
         let { limit = 5, skip = 0 } = req.query;
@@ -19,6 +22,14 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
     try {
+        let policy = policyFor(req.user);
+
+        if(!policy.can('create', 'Category')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         let payload = req.body;
         let category = new Model(payload);
         await category.save(); 
@@ -38,6 +49,14 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
     try {
+        let policy = policyFor(req.user);
+
+        if(!policy.can('update', 'Category')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         let payload = req.body;
         let category = await Model.findOneAndUpdate({_id: req.params.id}, payload, {new: true, runValidators: true});
         return res.json(category);
@@ -55,6 +74,14 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next){
     try {
+        let policy = policyFor(req.user);
+
+        if(!policy.can('delete', 'Category')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         let category = await Model.findOneAndDelete({_id: req.params.id});
         return res.json(category);
     } catch(err) {
