+++ app/product/controller.js
@@ -1,26 +1,75 @@
-// import model `Product`
+// import model
 const Model = require('./model');
+const Category = require('../category/model');
+const Tag = require('../tag/model');
 const config = require('../config');
 
 // Package Upload File
 const fs = require('fs');
 const path = require('path');
 
+// Hak Akses
+const { policyFor } = require('../policy');
+
 async function index(req, res, next) {
     try {
-        let { limit = 5, skip = 0 } = req.query;
-        let products = await Model.find().limit(parseInt(limit)).skip(parseInt(skip));
-        return res.json(products);
+        let { limit = 10, skip = 0, q = '', category = '', tags = [] } = req.query;
+        let criteria = {};
+        if(q.length){
+             // --- gabungkan dengan criteria --- //
+            criteria = {...criteria, name: {$regex: `${q}`, $options: 'i'}}
+        }
+        if(category.length){
+            category = await Category.findOne({name: {$regex: `${category}`}, $options: 'i'});
+            if(category) {
+                criteria = {...criteria, category: category._id}
+            }
+        }
+        if(tags.length){
+            tags = await Tag.find({name: {$in: tags}});
+            criteria = {...criteria, tags: {$in: tags.map(tag => tag._id)}}
+        }
+        let count = await Model.find(criteria).countDocuments();      
+        let products = await Model.find(criteria)
+            .populate('category')
+            .populate('tags')
+            .limit(parseInt(limit))
+            .skip(parseInt(skip))
+            .select('-__v');
+        return res.json({data: products, count});
     } catch (error) {
         next(error);
     }
-
 }
 
 async function store(req, res, next) {
+
     try {
+        let policy = policyFor(req.user);
+
+        if(!policy.can('create', 'Product')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         let payload = req.body;
 
+        if(payload.category){
+            let category = await Category.findOne({name: {$regex: payload.category, $options: 'i' }});
+            if(category) {
+                payload = {...payload, category: category._id}; 
+            } else {
+                delete payload.category; 
+            }
+        }
+        if(payload.tags && payload.tags.length){
+            let tags = await Tag.find({name: {$in: payload.tags}});
+            if(tags.length){
+                // jika ada, maka kita ambil `_id` untuk masing-masing `Tag` dan gabungkan dengan payload
+                payload = {...payload, tags: tags.map( tag => tag._id)}
+            }
+        }        
         if (req.file) {
             let tmp_path = req.file.path;
             let originalExt = req.file.originalname.split('.')[req.file.originalname.split('.').length - 1];
@@ -74,8 +123,30 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
     try {
+        //--- cek policy ---/
+        let policy = policyFor(req.user);
+        if(!policy.can('update', 'Product')){ 
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk mengupdate produk`
+            });
+        }
         let payload = req.body;
-
+        if(payload.category){
+            let category = await Category.findOne({name: {$regex: payload.category, $options: 'i'}});
+            if(category){ 
+                payload = {...payload, category: category._id};
+            } else {
+                delete payload.category;
+            }
+        }
+        if(payload.tags && payload.tags.length){
+            let tags = await Tag.find({name: {$in: payload.tags}});
+            if(tags.length){
+                // jika ada, maka kita ambil `_id` untuk masing-masing `Tag` dan gabungkan dengan payload
+                payload = {...payload, tags: tags.map( tag => tag._id)}
+            }
+        }
         if (req.file) {
             let tmp_path = req.file.path;
             let originalExt = req.file.originalname.split('.')[req.file.originalname.split('.').length - 1];
@@ -91,10 +162,8 @@ async function update(req, res, next) {
                 try {
                     // Find Produk yang akan di update
                     let product = await Model.findOne({_id: req.params.id});
-
                     // dapatkan path lengkap ke tempat penyimpanan image berdasarkan `product.image_url`
                     let currentImage = `${config.rootPath}/public/upload/${product.image_url}`;
-
                     // cek apakah `file` benar-benar ada di file system
                     if(fs.existsSync(currentImage)){
                         // hapus jika ada.
@@ -103,7 +172,6 @@ async function update(req, res, next) {
                     // Update
                     product = await Model.findOneAndUpdate({_id: req.params.id}, {...payload, image_url: filename}, {new: true, runValidators: true});
                     return res.json(product);
-
                 } catch (error) {
                     // jika error, hapus file yang sudah terupload pada direktori
                     fs.unlinkSync(target_path);
@@ -123,8 +191,8 @@ async function update(req, res, next) {
                 next(error);
             });
         } else {
-            let product = new Model(payload);
-            await product.findOneAndUpdate({_id: req.params.id}, payload, {new: true, runValidators: true});
+            // let product = new Model(payload);
+            let product = await Model.findOneAndUpdate({_id: req.params.id}, payload, {new: true, runValidators: true});
             return res.json(product);
         }
     } catch (error) {
@@ -141,16 +209,22 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next){
     try {
+        //--- cek policy ---/
+        let policy = policyFor(req.user);
+        if(!policy.can('delete', 'Product')){ 
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk menghapus produk`
+            });
+        }
         let product = await Model.findOneAndDelete({_id: req.params.id});
         let currentImage = `${config.rootPath}/public/upload/${product.image_url}`;
-
         if(fs.existsSync(currentImage)){
             fs.unlinkSync(currentImage)
         }
-
         return res.json(product);
-    } catch(err) {
-        next(err);
+    } catch(error) {
+        next(error);
     }
 }
